pyproject.toml
[project]
name = "nexis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "litellm>=1.83.8",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.7.23",
]
